remove button from tutorProfileCard. this card only shows info
