+++ src/components/cards/tutorProfileCard.js
@@ -1,8 +1,7 @@
-import React, { useEffect } from 'react'
+import React from 'react'
 import {
   Avatar,
   Box,
-  Button,
   Card,
   Divider,
   Grid,
@@ -10,11 +9,8 @@ import {
   Typography
 } from '@material-ui/core'
 
-import { getTutorSelectedInfo } from 'src/redux/actions/student/student_publications'
-import { connect } from 'react-redux'
 import { Rating } from '@material-ui/lab'
 import QualificationCard from 'src/components/cards/QualificationCard'
-import ProfileViewSkeleton from 'src/components/skeletons/ProfileViewSkeleton'
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -52,123 +48,88 @@ const useStyles = makeStyles((theme) => ({
 
 const ProfileView = (props) => {
   const classes = useStyles()
-  const { idTutor, getTutorSelectedInfo, tutor, loading, next } = props
-  useEffect(
-    () => {
-      getTutorSelectedInfo(idTutor)
-    },
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    []
-  )
-  const handleClick = () => {
-    next(tutor)
-  }
+  const { tutor } = props
   return (
     <>
       <Card className={classes.root}>
         <Box display="flex" flexDirection="column" justifyContent="center">
-          {loading ? (
-            <ProfileViewSkeleton />
-          ) : (
-            <>
-              <Box className={classes.title} textAlign="center">
-                <Typography variant="h3">
-                  <b>
-                    {tutor.user.first_name} {tutor.user.last_name}
-                  </b>
+          <Box className={classes.title} textAlign="center">
+            <Typography variant="h3">
+              <b>
+                {tutor.user.first_name} {tutor.user.last_name}
+              </b>
+            </Typography>
+            <Typography variant="h4">Tutor</Typography>
+          </Box>
+          <Divider />
+          <Grid container>
+            <Grid item xs={8}>
+              <Box
+                className={classes.principalInformation}
+                display="flex"
+                flexDirection="column">
+                <Typography className={classes.contentPrincipal} variant="h4">
+                  {getAge(tutor.user.birthday)} años
                 </Typography>
-                <Typography variant="h4">Tutor</Typography>
+                <Typography className={classes.contentPrincipal} variant="h4">
+                  {tutor.skills}
+                </Typography>
+                <Rating value={4} size="large" readOnly />
+                <Typography className={classes.contentPrincipal} variant="h4">
+                  Promedio: 4.5 de 23 calificaciones
+                </Typography>
               </Box>
-              <Divider />
-              <Grid container>
-                <Grid item xs={8}>
-                  <Box
-                    className={classes.principalInformation}
-                    display="flex"
-                    flexDirection="column">
-                    <Typography
-                      className={classes.contentPrincipal}
-                      variant="h4">
-                      {getAge(tutor.user.birthday)} años
-                    </Typography>
-                    <Typography
-                      className={classes.contentPrincipal}
-                      variant="h4">
-                      {tutor.skills}
-                    </Typography>
-                    <Rating value={4} size="large" readOnly />
-                    <Typography
-                      className={classes.contentPrincipal}
-                      variant="h4">
-                      Promedio: 4.5 de 23 calificaciones
-                    </Typography>
-                  </Box>
-                </Grid>
-                <Grid item xs={4}>
-                  <Box display="flex" justifyContent="center">
-                    <Avatar
-                      className={classes.cover}
-                      alt="user photo"
-                      src="/static/images/avatars/avatar_6.png"
-                    />
-                  </Box>
-                </Grid>
-              </Grid>
-              <Divider />
-              <Box display="flex" flexDirection="row">
-                <Box className={classes.secondInformation}>
+            </Grid>
+            <Grid item xs={4}>
+              <Box display="flex" justifyContent="center">
+                <Avatar
+                  className={classes.cover}
+                  alt="user photo"
+                  src="/static/images/avatars/avatar_6.png"
+                />
+              </Box>
+            </Grid>
+          </Grid>
+          <Divider />
+          <Box display="flex" flexDirection="row">
+            <Box className={classes.secondInformation}>
+              <Box textAlign="center">
+                <Typography variant="h4">
+                  <b>Metodologia</b>
+                </Typography>
+              </Box>
+              <Box>
+                <Typography align="justify">{tutor.methodology}</Typography>
+              </Box>
+              {tutor.trajectory === '' ? (
+                <></>
+              ) : (
+                <>
                   <Box textAlign="center">
                     <Typography variant="h4">
-                      <b>Metodologia</b>
+                      <b>Experiencia</b>
                     </Typography>
                   </Box>
                   <Box>
-                    <Typography align="justify">{tutor.methodology}</Typography>
+                    <Typography align="justify">{tutor.trajectory}</Typography>
                   </Box>
-                  {tutor.trajectory === '' ? (
-                    <></>
-                  ) : (
-                    <>
-                      <Box textAlign="center">
-                        <Typography variant="h4">
-                          <b>Experiencia</b>
-                        </Typography>
-                      </Box>
-                      <Box>
-                        <Typography align="justify">
-                          {tutor.trajectory}
-                        </Typography>
-                      </Box>
-                    </>
-                  )}
-                </Box>
-                <Box className={classes.divider}>
-                  <Divider orientation="vertical" />
-                </Box>
-                <Box className={classes.qualifications}>
-                  <Box textAlign="center">
-                    <Typography variant="h4">
-                      <b>Reseñas</b>
-                    </Typography>
-                  </Box>
-                  <QualificationCard />
-                </Box>
+                </>
+              )}
+            </Box>
+            <Box className={classes.divider}>
+              <Divider orientation="vertical" />
+            </Box>
+            <Box className={classes.qualifications}>
+              <Box textAlign="center">
+                <Typography variant="h4">
+                  <b>Reseñas</b>
+                </Typography>
               </Box>
-            </>
-          )}
+              <QualificationCard />
+            </Box>
+          </Box>
         </Box>
       </Card>
-
-      <Box className={classes.nextButton}>
-        <Button
-          size="large"
-          variant="contained"
-          color="primary"
-          onClick={handleClick}
-          className={classes.button}>
-          Siguiente
-        </Button>
-      </Box>
     </>
   )
 }
@@ -186,11 +147,4 @@ function getAge(fecha) {
   return edad
 }
 
-const mapStateToProps = (state) => ({
-  tutor: state.publications.tutorInfo,
-  loading: state.publications.loadingTutor
-})
-
-export default connect(mapStateToProps, {
-  getTutorSelectedInfo
-})(ProfileView)
+export default ProfileView
